Move all the code to the controller
Since we can't reasonably want to test the actual endpoints for the router, extract the methods and make a controller to hold all the real logic where it can be tested in isolation from the router.
Also add the version number from the package.json to the emitted response to the hello endpoint.
